moving towards offline capability

+++ client/src/store/actions.js
@@ -5,6 +5,7 @@ import {
 } from './mutations';
 
 const appService = require('./service');
+const storageService = require('./storageService');
 const {navigateTo} = require('../utils/router');
 
 export const constants = {
@@ -22,100 +23,111 @@ export const constants = {
     UPDATE_ACCOUNT_CONFIG:'UPDATE_ACCOUNT_CONFIG',
 }
 
-function actionCreator(response, goodRequest, badRequest){
-    if(response.status === 200){
-        goodRequest(response);
-    }
-    else if(response.status === 401){
-        localStorage.clear('jwt');
-        navigateTo('/');
-    }
-    else{
-        if(badRequest){
-            badRequest(response);
+async function actionCreator(onlineFunc, goodRequest, badRequest, offlineFunc){
+    console.log(navigator.onLine)
+    // If user is connected to a network they communicate with the server directly, no need to mess around with local storage
+    if(navigator.onLine){
+        const response = await onlineFunc();
+        if(response.status === 200){
+            goodRequest(response);
+        }
+        else if(response.status === 401){
+            localStorage.clear('jwt');
+            navigateTo('/');
+        }
+        else{
+            if(badRequest){
+                badRequest(response);
+            }
         }
     }
+
+    // // If user is offline, deal directly with localStorage
+    // if(!navigator.onLine && offlineFunc){
+    //     offlineFunc();
+    // }
+    
 }
 export default {
     
     
     async [constants.RETRIEVE_CHECKLISTS](context){
         actionCreator(
-            await appService.retrieveChecklists(),
+            () => appService.retrieveChecklists(),
             (response) => context.commit(SET_CHECKLISTS, response.results),
-            (response) => console.log('ERROR ', response)
+            (error) => console.log('ERROR ', error),
+            
+            () => storageService.retrieveChecklists()
         )
     },
     async [constants.ADD_CHECKLIST](context, title){
-        // if checklist is successfully added to database, push the added item to checklists in store
-            // message returns an array of responses, the first from inserting the checklist, the second an array of checklists from SELECT
-            // response.message[1][0] takes the first element out of the second array
-        
-        // if checklist isn't successfully added to database e.g. user is offline
-            // get userId from stored JWT
-            // assign a negative Id to checklist, pinned: false
-            // push checklist to localStorage array that will be synced with database once they go back online
-            // push checklist to store
         actionCreator(
-            await appService.addChecklist(title),
+            () => appService.addChecklist(title),
             (response) => context.commit(M_ADD_CHECKLIST, response.message[1][0]),
-            (response) => console.log('ERROR ', response)
+            (error) => console.log('ERROR ', error),
+
+            () => storageService.addChecklist(title)
         )
-        
     },
 
     async [constants.UPDATE_CHECKLIST](context, item){
         actionCreator(
-            await appService.updateChecklist(item),
-            (response) => context.commit(M_UPDATE_CHECKLIST, item)
+            () => appService.updateChecklist(item),
+            () => context.commit(M_UPDATE_CHECKLIST, item),
+            (error) => console.log('ERROR ', error),
+
+            () => context.commit(M_UPDATE_CHECKLIST, item)
         )
     },
 
     async [constants.DELETE_CHECKLIST](context, id){
         actionCreator(
-            await appService.deleteChecklist(id),
-            (response) => context.commit(M_DELETE_CHECKLIST, id)
+            () => appService.deleteChecklist(id),
+            () => context.commit(M_DELETE_CHECKLIST, id),
+            (error) => console.log('ERROR ', error),
+
+            () => context.commit(M_DELETE_CHECKLIST, id)
         )
     },
 
     async [constants.RETRIEVE_CHECKLIST_ITEMS](context){
         actionCreator(
-            await appService.retrieveChecklistItems(),
+            () => appService.retrieveChecklistItems(),
             (response) => context.commit(SET_CHECKLIST_ITEMS, response.items)
         )
     },
 
     async [constants.ADD_CHECKLIST_ITEM](context, {name, checklistId}){
         actionCreator(
-            await appService.addChecklistItem(name, checklistId),
+            () => appService.addChecklistItem(name, checklistId),
             (response) => context.commit(M_ADD_CHECKLIST_ITEM, response.message[1][0])
         )
     },
 
     async [constants.UPDATE_CHECKLIST_ITEM](context, item){
         actionCreator(
-            await appService.updateChecklistItem(item.Id, {name:item.Name, checked:item.Checked}),
+            () => appService.updateChecklistItem(item.Id, {name:item.Name, checked:item.Checked}),
             (response) => context.commit(M_UPDATE_CHECKLIST_ITEM, item)
         )
     },
 
     async [constants.DELETE_CHECKLIST_ITEM](context, itemId){
         actionCreator(
-            await appService.deleteChecklistItem(itemId),
+            () => appService.deleteChecklistItem(itemId),
             (response) => context.commit(M_DELETE_CHECKLIST_ITEM, itemId)
         )
     },
 
     async [constants.RETRIEVE_ACCOUNT_CONFIG](context){
         actionCreator(
-            await appService.retrieveAccountConfig(),
+            () => appService.retrieveAccountConfig(),
             (response) => context.commit(SET_ACCOUNT_CONFIG, response.accountConfig[0])
         )
     },
 
     async [constants.UPDATE_ACCOUNT_CONFIG](context, config){
         actionCreator(
-            await appService.updateAccountConfig(accountConfigToNumber(config)),
+            () => appService.updateAccountConfig(accountConfigToNumber(config)),
             (response) => context.commit(M_UPDATE_ACCOUNT_CONFIG, config)
         )
     }
